logging: move testsuite from robotsports/proof_is_alive to falcons/test_mra_logger
