+++ components/falcons/test_mra_logger/dependencies
@@ -1,0 +1,1 @@
+src:test_helpers
