+++ components/falcons/test_mra_logger/src/BUILD
@@ -1,0 +1,15 @@
+cc_library(
+    name = "test_helpers",
+    hdrs = [
+        "test_helpers.hpp",
+    ],
+    srcs = [
+        "test_helpers.cpp",
+    ],
+    visibility = ["//visibility:public"],
+    deps = [
+        "//libraries/logging",
+        "@com_google_googletest//:gtest_main",
+    ],
+)
+
