--- components/robotsports/proof_is_alive/test/BUILD
@@ -1,13 +1,0 @@
-cc_test(
-    name = "test_mra_logger",
-    srcs = [
-        "test_mra_logger.cpp",
-        "test_mra_logger_helpers.hpp",
-        "test_mra_logger_helpers.cpp",
-    ],
-    deps = [
-        "//components/robotsports/proof_is_alive:implementation",
-        "@com_google_googletest//:gtest_main",
-    ],
-)
-
